+++ src/hash.ts
@@ -1,0 +1,9 @@
+/**
+ * Create a unique hash of the system information.
+ * @param {info object} info
+*  @returns {string} hash
+*  @example
+* 	const hash = createHash(info);
+*  	console.log(hash);
+*  	=> 'a1b2c3d4e5f6g7h8i9j0'
+*/